+++ Database/ItemTypeConfiguration.cs
@@ -30,7 +30,7 @@ internal sealed class ItemTypeConfiguration : IEntityTypeConfiguration<Item>
 
             entityTypeBuilder
                 .Property(nameof(Item.QualityLevel))
-                .HasColumnType("tinyint") // Zdefiniowanie typu komórki tinyint wartości od 0-255
+                .HasColumnType("smallint") // Zdefiniowanie typu komórki tinyint wartości od 0-255
                 .HasColumnName(nameof(Item.QualityLevel))
                 .IsRequired();
 
@@ -41,7 +41,7 @@ internal sealed class ItemTypeConfiguration : IEntityTypeConfiguration<Item>
 
             entityTypeBuilder
                 .Property(nameof(Item.State))
-                .HasColumnType("tinyint")
+                .HasColumnType("smallint")
                 .HasColumnName(nameof(Item.State))
                 .IsRequired();
